src/scraper/requirements.txt
GitPython==3.1.43
python-dotenv==1.0.1